requirements.txt
fastapi>=0.110,<0.117
uvicorn[standard]>=0.23,<0.36
httpx>=0.24,<0.29
python-dotenv>=1.0,<2.0
pydantic>=2.6,<3
python-multipart>=0.0.6,<0.0.21
email-validator>=2.0,<3.0
PyJWT>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.0.0,<5.0.0
# MySQL Database Dependencies
sqlalchemy>=2.0.0,<3.0.0
mysql-connector-python>=8.0.0,<9.0.0
pymysql>=1.1.0,<2.0.0
cryptography>=41.0.0
# Cloudflare R2 support
boto3>=1.26.0,<4.0.0
